sleep-remap-powerkey/pyproject.toml
[project]
name = "sleep-remap-powerkey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jacbart", email = "jacbart@gmail.com" }
]
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
linux = ["python-uinput>=0.11.2"]

[project.scripts]
sleep-remap-powerkey = "sleep_remap_powerkey:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
